+++ NEWS.md
@@ -1,6 +1,7 @@
 # AzureContainers 1.0.1.9000
 
 * Ensure dir for Kubernetes config file exists before writing the file
+* Add `wait` argument to `create_aci` and `create_aks` methods; rely on AzureRMR 2.0 for implementation
 
 # AzureContainers 1.0.1